# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workout Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new trainer (creates organization)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful, verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "409":
          description: Username or email already taken
  /auth/verify-email:
    get:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email address with token
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verified and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "404":
          description: Invalid token
        "410":
          description: Token expired
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend email verification link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationBody"
      responses:
        "200":
          description: Verification email sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "429":
          description: Rate limited
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
        "403":
          description: Email not verified
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List all exercises
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create an exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseBody"
      responses:
        "201":
          description: Created exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get an exercise by ID
      parameters:
        - $ref: "#/components/parameters/ExerciseId"
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
    put:
      operationId: updateExercise
      tags: [exercises]
      summary: Update an exercise
      parameters:
        - $ref: "#/components/parameters/ExerciseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExerciseBody"
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete an exercise
      parameters:
        - $ref: "#/components/parameters/ExerciseId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
...
    ExerciseId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        email:
          type: string
      required:
        - message
        - email
    ResendVerificationBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
        organizationName:
          type: string
      required:
        - name
        - email
        - username
        - password
        - organizationName
    LoginBody:
      type: object
      properties:
        login:
          type: string
          description: Username or email
        password:
          type: string
      required:
        - login
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          type: string
          enum: [trainer, client]
        organizationId:
          type: integer
        organizationName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - organizationId
        - organizationName
        - createdAt
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        measurementType:
          type: string
          enum: [reps, seconds, meters]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - measurementType
        - createdAt
    CreateExerciseBody:
      type: object
      properties:
        name:
          type: string
        measurementType:
          type: string
          enum: [reps, seconds, meters]
      required:
        - name
        - measurementType
    UpdateExerciseBody:
      type: object
      properties:
        name:
          type: string
        measurementType:
          type: string
          enum: [reps, seconds, meters]
      required:
        - name
        - measurementType
...